+++ mods/Veil/common/src/main/java/foundry/veil/impl/client/render/pipeline/PatchStateShard.java
@@ -1,6 +1,7 @@
 package foundry.veil.impl.client.render.pipeline;
 
 import foundry.veil.Veil;
+import foundry.veil.api.client.render.VeilRenderSystem;
 import net.minecraft.client.renderer.RenderStateShard;
 import org.jetbrains.annotations.ApiStatus;
 
@@ -13,7 +14,15 @@ public class PatchStateShard extends RenderStateShard {
     private final int patchVertices;
 
     public PatchStateShard(int patchVertices) {
-        super(Veil.MODID + ":patches", () -> glPatchParameteri(GL_PATCH_VERTICES, patchVertices), () -> glPatchParameteri(GL_PATCH_VERTICES, 1));
+        super(Veil.MODID + ":patches", () -> {
+            if (VeilRenderSystem.tessellationSupported()) {
+                glPatchParameteri(GL_PATCH_VERTICES, patchVertices);
+            }
+        }, () -> {
+            if (VeilRenderSystem.tessellationSupported()) {
+                glPatchParameteri(GL_PATCH_VERTICES, 1);
+            }
+        });
         this.patchVertices = patchVertices;
     }